feat(search): replace legacy SearchBox by SearchBoxRework
