+++ src/features/search/components/SearchBox.tsx
@@ -1,60 +1,51 @@
 import { t } from '@lingui/macro'
-import { useFocusEffect, useNavigation } from '@react-navigation/native'
+import { useNavigation } from '@react-navigation/native'
 import React, { useCallback, useState } from 'react'
-import {
-  Keyboard,
-  NativeSyntheticEvent,
-  Platform,
-  TextInputSubmitEditingEventData,
-} from 'react-native'
+import { NativeSyntheticEvent, Platform, TextInputSubmitEditingEventData } from 'react-native'
 import styled from 'styled-components/native'
 import { v4 as uuidv4 } from 'uuid'
 
 import { UseNavigationType } from 'features/navigation/RootNavigator'
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
+import { useLocationChoice } from 'features/search/components/locationChoice.utils'
+import { LocationType } from 'features/search/enums'
 import { useSearch, useStagedSearch } from 'features/search/pages/SearchWrapper'
-import { accessibilityAndTestId } from 'libs/accessibilityAndTestId'
+import { useShowResults } from 'features/search/pages/useShowResults'
 import { analytics } from 'libs/firebase/analytics'
 import { HiddenAccessibleText } from 'ui/components/HiddenAccessibleText'
 import { SearchInput } from 'ui/components/inputs/SearchInput'
-import { Touchable } from 'ui/components/touchable/Touchable'
+import { TouchableOpacity } from 'ui/components/TouchableOpacity'
 import { ArrowPrevious as DefaultArrowPrevious } from 'ui/svg/icons/ArrowPrevious'
 import { MagnifyingGlass as DefaultMagnifyingGlass } from 'ui/svg/icons/MagnifyingGlass'
+import { getSpacing } from 'ui/theme'
 import { getHeadingAttrs } from 'ui/theme/typographyAttrs/getHeadingAttrs'
-
-const LeftIcon: React.FC<{ onPressArrowBack: () => void }> = ({ onPressArrowBack }) => {
-  const { searchState } = useSearch()
-  if (searchState.showResults)
-    return (
-      <Touchable onPress={onPressArrowBack} {...accessibilityAndTestId(t`Revenir en arrière`)}>
-        <ArrowPrevious />
-      </Touchable>
-    )
-  return <MagnifyingGlass />
-}
 
 type Props = {
   searchInputID: string
+  showLocationButton?: boolean
+  onFocusState?: (focus: boolean) => void
+  isFocus?: boolean
+  accessibleHiddenTitle?: string
 }
 
-export const SearchBox: React.FC<Props> = ({ searchInputID }) => {
+export const SearchBox: React.FC<Props> = ({
+  searchInputID,
+  showLocationButton,
+  onFocusState,
+  isFocus,
+  accessibleHiddenTitle,
+}) => {
   const { navigate } = useNavigation<UseNavigationType>()
-  const { searchState, dispatch } = useSearch()
   const { searchState: stagedSearchState, dispatch: stagedDispatch } = useStagedSearch()
-  const [query, _setQuery] = useState<string>('')
+  const { searchState, dispatch } = useSearch()
+  const [query, setQuery] = useState<string>(searchState.query || '')
   const accessibilityDescribedBy = uuidv4()
-
-  function setQuery(value: string) {
-    stagedDispatch({ type: 'SET_QUERY', payload: value })
-    _setQuery(value)
-  }
-
-  useFocusEffect(
-    useCallback(() => {
-      setQuery(searchState.query)
-      // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [searchState.query])
-  )
+  const showResults = useShowResults()
+  const { locationFilter } = stagedSearchState
+  const { locationType } = locationFilter
+  // PLACE and VENUE belong to the same section
+  const section = locationType === LocationType.VENUE ? LocationType.PLACE : locationType
+  const { label: locationLabel } = useLocationChoice(section)
 
   const resetSearch = () => {
     navigate(...getTabNavConfig('Search', { query: '' }))
@@ -63,11 +54,16 @@ export const SearchBox: React.FC<Props> = ({ searchInputID }) => {
 
   const onPressArrowBack = () => {
     setQuery('')
+    if (onFocusState) onFocusState(false)
+    stagedDispatch({ type: 'SET_QUERY', payload: '' })
     dispatch({ type: 'SET_QUERY', payload: '' })
     dispatch({ type: 'SHOW_RESULTS', payload: false })
     dispatch({ type: 'INIT' })
-    Keyboard.dismiss()
   }
+
+  const onPressLocationButton = useCallback(() => {
+    navigate('LocationFilter')
+  }, [navigate])
 
   const onSubmitQuery = (event: NativeSyntheticEvent<TextInputSubmitEditingEventData>) => {
     const queryText = event.nativeEvent.text
@@ -76,7 +72,7 @@ export const SearchBox: React.FC<Props> = ({ searchInputID }) => {
     // these are the two potentially 'staged' filters that we want to commit to the global search state.
     // We also want to commit the price filter, as beneficiary users may have access to different offer
     // price range depending on their available credit.
-    const { locationFilter, offerCategories, priceRange } = stagedSearchState
+    const { offerCategories, priceRange } = stagedSearchState
     navigate(
       ...getTabNavConfig('Search', {
         showResults: true,
@@ -91,40 +87,62 @@ export const SearchBox: React.FC<Props> = ({ searchInputID }) => {
 
   return (
     <React.Fragment>
-      <HiddenAccessibleText {...getHeadingAttrs(1)}>
-        {t`Recherche une offre, un titre, un lieu...`}
-      </HiddenAccessibleText>
-      <StyledSearchInput
-        searchInputID={searchInputID}
-        value={query}
-        onChangeText={setQuery}
-        placeholder={t`Offre, artiste...`}
-        autoFocus={false}
-        inputHeight="regular"
-        LeftIcon={() => <LeftIcon onPressArrowBack={onPressArrowBack} />}
-        onSubmitEditing={onSubmitQuery}
-        accessibilityLabel={t`Rechercher un artiste, titre, lieu...`}
-        onPressRightIcon={resetSearch}
-        accessibilityDescribedBy={accessibilityDescribedBy}
-      />
+      {!!accessibleHiddenTitle && (
+        <HiddenAccessibleText {...getHeadingAttrs(1)}>{accessibleHiddenTitle}</HiddenAccessibleText>
+      )}
+      <SearchInputContainer marginHorizontal={showResults || isFocus ? getSpacing(6) : 0}>
+        {showResults || isFocus ? (
+          <StyledTouchableOpacity testID="previousButton" onPress={onPressArrowBack}>
+            <ArrowPrevious />
+          </StyledTouchableOpacity>
+        ) : null}
+        <StyledSearchInput
+          searchInputID={searchInputID}
+          value={query}
+          onChangeText={setQuery}
+          placeholder={t`Offre, artiste...`}
+          autoFocus={isFocus}
+          inputHeight="regular"
+          LeftIcon={() => <MagnifyingGlassIcon />}
+          onSubmitEditing={onSubmitQuery}
+          onPressRightIcon={resetSearch}
+          onFocusState={onFocusState}
+          testID="searchInput"
+          onPressLocationButton={showLocationButton ? onPressLocationButton : undefined}
+          locationLabel={locationLabel}
+        />
+      </SearchInputContainer>
       <HiddenAccessibleText nativeID={accessibilityDescribedBy}>
         {t`Indique le nom d'une offre ou d'un lieu puis lance la recherche à l'aide de la touche
-        "Entrée"`}
+          "Entrée"`}
       </HiddenAccessibleText>
     </React.Fragment>
   )
 }
 
-const StyledSearchInput = styled((props) => <SearchInput {...props} />).attrs(({ theme }) => ({
-  focusOutlineColor: theme.colors.black,
-}))({
+const StyledSearchInput = styled((props) => <SearchInput {...props} />)({
   outlineOffset: 0,
 })
 
 const ArrowPrevious = styled(DefaultArrowPrevious).attrs(({ theme }) => ({
-  size: theme.icons.sizes.smaller,
+  size: theme.icons.sizes.small,
 }))``
 
 const MagnifyingGlass = styled(DefaultMagnifyingGlass).attrs(({ theme }) => ({
   size: theme.icons.sizes.smaller,
 }))``
+
+const MagnifyingGlassIcon = styled(MagnifyingGlass).attrs(({ theme }) => ({
+  size: theme.icons.sizes.smaller,
+}))``
+
+const SearchInputContainer = styled.View<{ marginHorizontal: number }>(({ marginHorizontal }) => ({
+  flexDirection: 'row',
+  alignItems: 'center',
+  justifyContent: 'center',
+  marginHorizontal,
+}))
+
+const StyledTouchableOpacity = styled(TouchableOpacity)({
+  marginRight: getSpacing(4),
+})
